requirements.txt
matplotlib==3.7.1
seaborn==0.13.1
pandas==2.0.3
scikit-learn==1.2.2
dill==0.3.4
scikeras == 0.13.0
keras == 3.3.3
tensorflow == 2.16.1
dash == 2.17.0
psycopg==3.1.18
python-dotenv==1.0.1